♻️  [Refactor]: Move Feature of Header from Layout.template.tsx to Header.tsx
- header와 관련된 기능을 Header.tsx로 옮겨 책임 분리를 진행 함.
